Load the attendance form in the same page and without caching
Load the attendance form in the same page and without caching so that the Back button will take the teacher back to the course page.

Submit the attendance in the same page and without caching

Submit the attendance in the same page and without caching so that the Back button will take the teacher back to the course page afterwards.